pyproject.toml
[project]
name = "ai-review-analysis-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
]
